requirements.txt

discord.py>=2.3.2
python-dotenv>=1.0.0
crewai>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
ddgs
asyncio>=3.4.3
langchain-community
litellm
openai>=1.0.0
ollama
langchain-ollama